+++ agent/src/agent.ts
@@ -210,8 +210,24 @@ export function errorToCodyError(error?: Error): CodyError | undefined {
 }
 
 export class Agent extends MessageHandler implements ExtensionClient {
+    // Used to track background work of the extension, like tree-sitter parsing.
+    // In several places in the extension, we register event handler that run
+    // background work (`Promise<void>` that we don't await on). We sometimes
+    // need to await on these promises, for example when writing deterministic
+    // tests.
+    private pendingPromises = new Set<Promise<any>>()
     public codeLenses = new AgentCodeLenses()
     public workspace = new AgentWorkspaceDocuments({
+        doPanic: (message: string) => {
+            const panicMessage =
+                '!PANIC! Client document content is out of sync with server document content'
+            process.stderr.write(panicMessage)
+            process.stderr.write(message + '\n')
+            this.notify('debug/message', {
+                channel: 'Document Sync Check',
+                message: panicMessage + '\n' + message,
+            })
+        },
         edit: (uri, callback, options) => {
             if (this.clientInfo?.capabilities?.edit !== 'enabled') {
                 logDebug('CodyAgent', 'client does not support operation: textDocument/edit')
@@ -388,12 +404,14 @@ export class Agent extends MessageHandler implements ExtensionClient {
             this.workspace.setActiveTextEditor(
                 this.workspace.newTextEditor(this.workspace.loadDocument(documentWithUri))
             )
+            this.pushPendingPromise(this.workspace.fireVisibleTextEditorsDidChange())
         })
 
         this.registerNotification('textDocument/didOpen', document => {
             const documentWithUri = ProtocolTextDocumentWithUri.fromDocument(document)
             const textDocument = this.workspace.loadDocument(documentWithUri)
             vscode_shim.onDidOpenTextDocument.fire(textDocument)
+            this.pushPendingPromise(this.workspace.fireVisibleTextEditorsDidChange())
             this.workspace.setActiveTextEditor(this.workspace.newTextEditor(textDocument))
         })
 
@@ -405,11 +423,13 @@ export class Agent extends MessageHandler implements ExtensionClient {
             this.workspace.setActiveTextEditor(textEditor)
 
             if (contentChanges.length > 0) {
-                vscode_shim.onDidChangeTextDocument.fire({
-                    document: textDocument,
-                    contentChanges,
-                    reason: undefined,
-                })
+                this.pushPendingPromise(
+                    vscode_shim.onDidChangeTextDocument.cody_fireAsync({
+                        document: textDocument,
+                        contentChanges,
+                        reason: undefined,
+                    })
+                )
             }
 
             if (document.selection) {
@@ -428,6 +448,7 @@ export class Agent extends MessageHandler implements ExtensionClient {
                 this.workspace.deleteDocument(documentWithUri.uri)
                 vscode_shim.onDidCloseTextDocument.fire(oldDocument)
             }
+            this.pushPendingPromise(this.workspace.fireVisibleTextEditorsDidChange())
         })
 
         this.registerNotification('textDocument/didSave', async params => {
@@ -460,6 +481,24 @@ export class Agent extends MessageHandler implements ExtensionClient {
             }
         })
 
+        this.registerAuthenticatedRequest('testing/awaitPendingPromises', async () => {
+            if (!vscode_shim.isTesting) {
+                throw new Error(
+                    'testing/awaitPendingPromises can only be called from tests. ' +
+                        'To fix this problem, set the environment variable CODY_SHIM_TESTING=true.'
+                )
+            }
+            await Promise.all(this.pendingPromises.values())
+            return null
+        })
+
+        this.registerAuthenticatedRequest('testing/memoryUsage', async () => {
+            if (!global.gc) {
+                throw new Error('testing/memoryUsage requires running node with --expose-gc')
+            }
+            global.gc()
+            return { usage: process.memoryUsage() }
+        })
         this.registerAuthenticatedRequest('testing/networkRequests', async () => {
             const requests = this.params.networkRequests ?? []
             return {
@@ -1035,6 +1074,14 @@ export class Agent extends MessageHandler implements ExtensionClient {
             contextFiltersProvider.setTestingContextFilters(contextFilters)
             return null
         })
+    }
+
+    private pushPendingPromise(pendingPromise: Promise<unknown>): void {
+        if (!vscode_shim.isTesting) {
+            return
+        }
+        this.pendingPromises.add(pendingPromise)
+        pendingPromise.finally(() => this.pendingPromises.delete(pendingPromise))
     }
 
     // ExtensionClient callbacks.