+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -27,6 +27,7 @@ import {
     Typewriter,
     allMentionProvidersMetadata,
     hydrateAfterPostMessage,
+    isAbortError,
     isDefined,
     isError,
     isFileURI,
@@ -266,6 +267,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                     message.contextFiles ?? [],
                     message.editorState,
                     message.addEnhancedContext ?? false,
+                    this.startNewSubmitOrEditOperation(),
                     'chat'
                 )
                 break
@@ -455,6 +457,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         userContextFiles: ContextItem[],
         editorState: ChatMessage['editorState'],
         addEnhancedContext: boolean,
+        abortSignal: AbortSignal,
         source?: EventSource,
         command?: DefaultChatCommands
     ): Promise<void> {
@@ -502,21 +505,25 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 if (submitType === 'user-newchat' && !this.chatModel.isEmpty()) {
                     span.addEvent('clearAndRestartSession')
                     await this.clearAndRestartSession()
+                    abortSignal.throwIfAborted()
                 }
 
                 this.chatModel.addHumanMessage({ text: inputText, editorState })
                 await this.saveSession()
+                abortSignal.throwIfAborted()
 
                 this.postEmptyMessageInProgress()
 
                 // Add user's current selection as context for chat messages.
                 const selectionContext = source === 'chat' ? await getContextFileFromSelection() : []
+                abortSignal.throwIfAborted()
 
                 const userContextItems: ContextItemWithContent[] = await resolveContextItems(
                     this.editor,
                     [...userContextFiles, ...selectionContext],
                     inputText
                 )
+                abortSignal.throwIfAborted()
 
                 span.setAttribute('strategy', this.config.useContext)
                 const prompter = new DefaultPrompter(
@@ -568,10 +575,18 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 }
 
                 try {
-                    const { prompt, tokenCounter } = await this.buildPrompt(prompter, sendTelemetry)
-                    this.streamAssistantResponse(requestID, prompt, span, firstTokenSpan)
+                    const { prompt, tokenCounter } = await this.buildPrompt(
+                        prompter,
+                        abortSignal,
+                        sendTelemetry
+                    )
+                    abortSignal.throwIfAborted()
+                    this.streamAssistantResponse(requestID, prompt, span, firstTokenSpan, abortSignal)
                     this.postRemainingTokensToWebview(tokenCounter)
                 } catch (error) {
+                    if (isAbortErrorOrSocketHangUp(error as Error)) {
+                        return
+                    }
                     if (isRateLimitError(error)) {
                         this.postError(error, 'transcript')
                     } else {
@@ -585,6 +600,19 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 }
             })
         })
+    }
+
+    private submitOrEditOperation: AbortController | undefined
+    public startNewSubmitOrEditOperation(): AbortSignal {
+        this.submitOrEditOperation?.abort()
+        this.submitOrEditOperation = new AbortController()
+        return this.submitOrEditOperation.signal
+    }
+    private cancelSubmitOrEditOperation(): void {
+        if (this.submitOrEditOperation) {
+            this.submitOrEditOperation.abort()
+            this.submitOrEditOperation = undefined
+        }
     }
 
     /**
@@ -602,6 +630,8 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         editorState: ChatMessage['editorState'],
         addEnhancedContext = true
     ): Promise<void> {
+        const abortSignal = this.startNewSubmitOrEditOperation()
+
         telemetryService.log('CodyVSCodeExtension:editChatButton:clicked', undefined, {
             hasV2Event: true,
         })
@@ -620,6 +650,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 contextFiles,
                 editorState,
                 addEnhancedContext,
+                abortSignal,
                 'chat'
             )
         } catch {
@@ -628,7 +659,8 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
     }
 
     private handleAbort(): void {
-        this.cancelInProgressCompletion()
+        this.cancelSubmitOrEditOperation()
+
         telemetryService.log('CodyVSCodeExtension:abortButton:clicked', { hasV2Event: true })
         telemetryRecorder.recordEvent('cody.sidebar.abortButton', 'clicked')
     }
@@ -933,12 +965,14 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
      */
     private async buildPrompt(
         prompter: DefaultPrompter,
+        abortSignal: AbortSignal,
         sendTelemetry?: (contextSummary: any, privateContextStats?: any) => void
     ): Promise<{ prompt: Message[]; tokenCounter: TokenCounter }> {
         const { promptInfo, tokenCounter } = await prompter.makePrompt(
             this.chatModel,
             this.authProvider.getAuthStatus().codyApiVersion
         )
+        abortSignal.throwIfAborted()
 
         // Update UI based on prompt construction
         // Includes the excluded context items to display in the UI
@@ -984,7 +1018,8 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         requestID: string,
         prompt: Message[],
         span: Span,
-        firstTokenSpan: Span
+        firstTokenSpan: Span,
+        abortSignal: AbortSignal
     ): void {
         logDebug('SimpleChatPanelProvider', 'streamAssistantResponse', {
             verbose: { requestID, prompt },
@@ -999,37 +1034,48 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             firstTokenSpan.end()
         }
 
+        abortSignal.throwIfAborted()
         this.postEmptyMessageInProgress()
-        this.sendLLMRequest(prompt, {
-            update: content => {
-                measureFirstToken()
-                span.addEvent('update')
-                this.postViewTranscript({
-                    speaker: 'assistant',
-                    text: PromptString.unsafe_fromLLMResponse(content),
-                    model: this.chatModel.modelID,
-                })
-            },
-            close: content => {
-                measureFirstToken()
-                recordExposedExperimentsToSpan(span)
-                span.end()
-                this.addBotMessage(requestID, PromptString.unsafe_fromLLMResponse(content))
-            },
-            error: (partialResponse, error) => {
-                if (!isAbortError(error)) {
+        this.sendLLMRequest(
+            prompt,
+            {
+                update: content => {
+                    abortSignal.throwIfAborted()
+                    measureFirstToken()
+                    span.addEvent('update')
+                    this.postViewTranscript({
+                        speaker: 'assistant',
+                        text: PromptString.unsafe_fromLLMResponse(content),
+                        model: this.chatModel.modelID,
+                    })
+                },
+                close: content => {
+                    abortSignal.throwIfAborted()
+                    measureFirstToken()
+                    recordExposedExperimentsToSpan(span)
+                    span.end()
+                    this.addBotMessage(requestID, PromptString.unsafe_fromLLMResponse(content))
+                },
+                error: (partialResponse, error) => {
+                    if (isAbortErrorOrSocketHangUp(error)) {
+                        throw error
+                    }
                     this.postError(error, 'transcript')
-                }
-                try {
-                    // We should still add the partial response if there was an error
-                    // This'd throw an error if one has already been added
-                    this.addBotMessage(requestID, PromptString.unsafe_fromLLMResponse(partialResponse))
-                } catch {
-                    console.error('Streaming Error', error)
-                }
-                recordErrorToSpan(span, error)
+                    try {
+                        // We should still add the partial response if there was an error
+                        // This'd throw an error if one has already been added
+                        this.addBotMessage(
+                            requestID,
+                            PromptString.unsafe_fromLLMResponse(partialResponse)
+                        )
+                    } catch {
+                        console.error('Streaming Error', error)
+                    }
+                    recordErrorToSpan(span, error)
+                },
             },
-        })
+            abortSignal
+        )
     }
 
     /**
@@ -1042,7 +1088,8 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             update: (response: string) => void
             close: (finalResponse: string) => void
             error: (completedResponse: string, error: Error) => void
-        }
+        },
+        abortSignal: AbortSignal
     ): Promise<void> {
         let lastContent = ''
         const typewriter = new Typewriter({
@@ -1058,9 +1105,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             },
         })
 
-        this.cancelInProgressCompletion()
-        const abortController = new AbortController()
-        this.completionCanceller = () => abortController.abort()
         try {
             const stream = this.chatClient.chat(
                 prompt,
@@ -1068,7 +1112,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                     model: this.chatModel.modelID,
                     maxTokensToSample: this.chatModel.contextWindow.output,
                 },
-                abortController.signal
+                abortSignal
             )
 
             for await (const message of stream) {
@@ -1078,32 +1122,19 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                         break
                     }
                     case 'complete': {
-                        this.completionCanceller = undefined
                         typewriter.close()
                         typewriter.stop()
                         break
                     }
                     case 'error': {
-                        this.cancelInProgressCompletion()
                         typewriter.close()
                         typewriter.stop(message.error)
                     }
                 }
             }
         } catch (error: unknown) {
-            if (!isAbortError(error as Error)) {
-                this.cancelInProgressCompletion()
-                typewriter.close()
-                typewriter.stop(error as Error)
-            }
-        }
-    }
-
-    private completionCanceller?: () => void
-    private cancelInProgressCompletion(): void {
-        if (this.completionCanceller) {
-            this.completionCanceller()
-            this.completionCanceller = undefined
+            typewriter.close()
+            typewriter.stop(isAbortErrorOrSocketHangUp(error as Error) ? undefined : (error as Error))
         }
     }
 
@@ -1179,7 +1210,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         if (!oldTranscript) {
             return this.newSession()
         }
-        this.cancelInProgressCompletion()
+        this.cancelSubmitOrEditOperation()
         const newModel = newChatModelFromSerializedChatTranscript(oldTranscript, this.chatModel.modelID)
         this.chatModel = newModel
 
@@ -1212,7 +1243,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             return
         }
 
-        this.cancelInProgressCompletion()
+        this.cancelSubmitOrEditOperation()
         await this.saveSession()
 
         this.chatModel = new SimpleChatModel(this.chatModel.modelID)
@@ -1310,7 +1341,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
 
         // Dispose panel when the panel is closed
         panel.onDidDispose(() => {
-            this.cancelInProgressCompletion()
+            this.cancelSubmitOrEditOperation()
             this._webviewPanel = undefined
             this._webview = undefined
             panel.dispose()
@@ -1401,6 +1432,6 @@ function newChatModelFromSerializedChatTranscript(
     )
 }
 
-function isAbortError(error: Error): boolean {
-    return error.message === 'aborted' || error.message === 'socket hang up'
+function isAbortErrorOrSocketHangUp(error: unknown): error is Error {
+    return Boolean(isAbortError(error) || (error && (error as any).message === 'socket hang up'))
 }