+++ vscode/webviews/Chat.tsx
@@ -1,6 +1,6 @@
 import { clsx } from 'clsx'
 import type React from 'react'
-import { type ReactNode, useCallback, useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
 import type {
     AuthStatus,
@@ -10,16 +10,16 @@ import type {
     TelemetryRecorder,
     TelemetryService,
 } from '@sourcegraph/cody-shared'
-import { Transcript } from './chat/Transcript'
+import { Transcript, focusLastHumanMessageEditor } from './chat/Transcript'
 import type { VSCodeWrapper } from './utils/VSCodeApi'
 
 import { truncateTextStart } from '@sourcegraph/cody-shared/src/prompt/truncation'
 import { CHAT_INPUT_TOKEN_BUDGET } from '@sourcegraph/cody-shared/src/token/constants'
 import styles from './Chat.module.css'
 import { TokenIndicators } from './components/RemainingTokens'
+import { ScrollDown } from './components/ScrollDown'
 
 interface ChatboxProps {
-    welcomeMessage?: ReactNode
     chatEnabled: boolean
     messageInProgress: ChatMessage | null
     transcript: ChatMessage[]
@@ -33,7 +33,6 @@ interface ChatboxProps {
 }
 
 export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>> = ({
-    welcomeMessage,
     messageInProgress,
     transcript,
     vscodeAPI,
@@ -148,16 +147,6 @@ export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>
 
     useEffect(() => {
         function handleKeyDown(event: KeyboardEvent) {
-            // Opt+> and Alt+> focus the last editor input, to make it easy for users to ask a followup
-            // question.
-            if (event.altKey && event.key === '>') {
-                event.preventDefault()
-                event.stopPropagation()
-                const allEditors = document.querySelectorAll<HTMLElement>('[data-lexical-editor="true"]')
-                const lastEditor = allEditors.item(allEditors.length - 1) as HTMLElement | undefined
-                lastEditor?.focus()
-            }
-
             // Esc to abort the message in progress.
             if (event.key === 'Escape' && messageInProgress) {
                 vscodeAPI.postMessage({ command: 'abort' })
@@ -188,7 +177,7 @@ export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>
             const focusElement = focusNode instanceof Element ? focusNode : focusNode?.parentElement
             const focusEditor = focusElement?.closest<HTMLElement>('[data-lexical-editor="true"]')
             if (focusEditor) {
-                focusEditor.focus()
+                focusEditor.focus({ preventScroll: true })
             }
         }
         window.addEventListener('focus', onFocus)
@@ -198,7 +187,7 @@ export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>
     }, [])
 
     return (
-        <div className={clsx(styles.container)}>
+        <div className={clsx(styles.container, 'tw-relative')}>
             {!chatEnabled && (
                 <div className={styles.chatDisabled}>
                     Cody chat is disabled by your Sourcegraph site administrator
@@ -206,7 +195,6 @@ export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>
             )}
             <Transcript
                 transcript={transcript}
-                welcomeMessage={welcomeMessage}
                 messageInProgress={messageInProgress}
                 feedbackButtonsOnSubmit={feedbackButtonsOnSubmit}
                 copyButtonOnSubmit={copyButtonOnSubmit}
@@ -218,6 +206,7 @@ export const Chat: React.FunctionComponent<React.PropsWithChildren<ChatboxProps>
                 postMessage={postMessage}
                 guardrails={guardrails}
             />
+            <ScrollDown onClick={focusLastHumanMessageEditor} />
             <div className={clsx(styles.tokenIndicators)}>
                 {remainingTokens && <TokenIndicators remainingTokens={remainingTokens} />}
             </div>